+++ vertx-sfdc-core/sampleconfig.json
