+++ vertx-sfdc-core/src/main/java/net/wissel/salesforce/vertx/Utils.java
