+++ vertx-sfdc-core/src/main/java/net/wissel/salesforce/vertx/auth/AuthInfoCodec.java
