+++ vertx-sfdc-core/src/main/java/net/wissel/salesforce/vertx/auth/Basic.java
