+++ vertx-sfdc-core/src/main/java/net/wissel/salesforce/vertx/config/ListenerConfig.java
