+++ vertx-sfdc-core/src/main/java/net/wissel/salesforce/vertx/listener/CometDChangeObject.java
