+++ vertx-sfdc-core/src/main/resources/logintemplate.xml
