+++ vertx-sfdc-core/src/test/java/net/wissel/salesforce/vertx/TestUtilities.java
