+++ vertx-sfdc-redis/src/main/java/net/wissel/salesforce/vertx/consumer/RedisDedup.java
